refactor(channels): hoist canonical CHANNEL_IDS into service-contracts
The assistant and gateway each maintained their own CHANNEL_IDS array,
free to drift apart silently. Move the canonical set into
@vellumai/service-contracts/channels as the single source of truth.

- The assistant adopts it wholesale: it re-exports CHANNEL_IDS /
  ChannelId / isChannelId and keeps the daemon-only parseChannelId
  convenience over the shared guard.
- The gateway keeps its narrower inbound list (no `platform`, whose
  absence the admission-policy routes depend on) but adds a
  `satisfies readonly CanonicalChannelId[]` guard so it can never list a
  channel the assistant doesn't recognize.

Also removes the gateway's unused parseChannelId.
